찜하기 예외처리 추가

+++ backend/src/main/java/com/reclebooks/recle/service/WishServiceImpl.java
@@ -7,6 +7,8 @@ import com.reclebooks.recle.repository.PostRepository;
 import com.reclebooks.recle.repository.UserRepository;
 import com.reclebooks.recle.repository.WishRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,10 +22,16 @@ public class WishServiceImpl implements WishService{
     private final UserRepository userRepository;
     private final PostRepository postRepository;
     @Override
-    public Long addWishList(Long postId, Long userId) {
+    public Long addWishList(Long postId, Long userId) throws Exception{
 
         User user = userRepository.findById(userId).get();
         Post post = postRepository.findById(postId).get();
+
+        for (Wish wish : user.getWishList()) {
+            if(wish.getPost().getId().equals(postId)){
+                throw new Exception("이미 찜함");
+            }
+        }
 
         Wish wish = new Wish();
         wish.setUser(user);
@@ -33,9 +41,13 @@ public class WishServiceImpl implements WishService{
     }
 
     @Override
-    public void deleteWishList(Long postId, Long userId) {
+    public void deleteWishList(Long postId, Long userId) throws Exception {
+         Wish wish = wishRepository.findByPostIdAndUserId(postId, userId).orElse(null);
 
-        Wish wish = wishRepository.findByPostIdAndUserId(postId,userId).orElse(null);
+         // 본인인지 체크
+         if (wish.getUser().getId() != userId) {
+             throw new Exception("찜하지 않았음");
+         }
         wishRepository.deleteById(wish.getId());
     }
 }
